VMS-13313: Unit test in progress
<==== Mercurial data ====>
Branch: x_vms_13313_refactor_archive_reindexing

+++ vms/server/nx_vms_server/src/test_support/mediaserver_launcher.cpp
@@ -123,7 +123,7 @@ void MediaServerLauncher::prepareToStart()
     m_configFile.close();
 
     m_mediaServerProcess.reset();
-    m_mediaServerProcess.reset(new MediaServerProcess(argv.size() - 1, (char**) argv.data()));
+    m_mediaServerProcess.reset(new MediaServerProcess((int) argv.size() - 1, (char**) argv.data()));
     connect(m_mediaServerProcess.get(), &MediaServerProcess::started, this,
         &MediaServerLauncher::started);
